server/requirements.txt
astroid==2.1.0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.2.0
coverage==4.5.2
Flask==1.0.2
Flask-JWT-Extended==3.17.0
idna==2.8
inflection==0.3.1
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
mysqlclient==1.3.14
nose==1.3.7
openapi-spec-validator==0.2.4
passlib==1.7.1
PyJWT==1.7.1
pylint==2.2.2
PyYAML==5.1
requests==2.21.0
six==1.12.0
SQLAlchemy==1.2.17
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.11.1
